+++ moban/jinja2/extensions.py
@@ -1,6 +1,6 @@
 from lml.plugin import PluginInfo, PluginManager
 
-import moban.constants as constants
+from moban import constants
 
 
 class PluginMixin:
@@ -58,12 +58,3 @@ class JinjaGlobalsManager(PluginManager, PluginMixin):
 def jinja_global(identifier, dict_obj):
     plugin = PluginInfo(constants.JINJA_GLOBALS_EXTENSION, tags=[identifier])
     plugin(dict_obj)
-
-
-class LibraryManager(PluginManager):
-    def __init__(self):
-        super(LibraryManager, self).__init__(constants.LIBRARY_EXTENSION)
-
-    def resource_path_of(self, library_name):
-        library = self.get_a_plugin(library_name)
-        return library.resources_path
