+++ tests/integration_tests/test_command_line_options.py
@@ -13,7 +13,7 @@ class TestCustomOptions:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
+            "moban.jinja2.engine.verify_the_existence_of_directories"
         )
         self.patcher2 = patch(
             "moban.engine_factory.verify_the_existence_of_directories"
@@ -21,7 +21,7 @@ class TestCustomOptions:
         self.patcher1.start()
         self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_custom_options(self, fake_template_doer):
         test_args = [
             "moban",
@@ -40,7 +40,7 @@ class TestCustomOptions:
                 "a.jj2", "config.yaml", "moban.output"
             )
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_minimal_options(self, fake_template_doer):
         test_args = ["moban", "-c", self.config_file, "-t", "a.jj2"]
         with patch.object(sys, "argv", test_args):
@@ -67,7 +67,7 @@ class TestOptions:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
+            "moban.jinja2.engine.verify_the_existence_of_directories"
         )
         self.patcher2 = patch(
             "moban.engine_factory.verify_the_existence_of_directories"
@@ -75,7 +75,7 @@ class TestOptions:
         self.patcher1.start()
         self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_default_options(self, fake_template_doer):
         test_args = ["moban", "-t", "a.jj2"]
         with patch.object(sys, "argv", test_args):
@@ -114,7 +114,7 @@ class TestNoOptions:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
+            "moban.jinja2.engine.verify_the_existence_of_directories"
         )
         self.patcher2 = patch(
             "moban.engine_factory.verify_the_existence_of_directories"
@@ -122,7 +122,7 @@ class TestNoOptions:
         self.patcher1.start()
         self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
@@ -136,7 +136,7 @@ class TestNoOptions:
                 ],
             )
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command_with_a_few_options(self, fake_template_doer):
         test_args = ["moban", "-t", "abc.jj2", "-o", "xyz.output"]
         with patch.object(sys, "argv", test_args):
@@ -151,7 +151,7 @@ class TestNoOptions:
                 ],
             )
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command_with_options(self, fake_template_doer):
         test_args = [
             "moban",
@@ -175,7 +175,7 @@ class TestNoOptions:
             )
 
     @raises(Exception)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command_without_output_option(self, fake_template_doer):
         test_args = ["moban", "-t", "abc.jj2"]
         with patch.object(sys, "argv", test_args):
@@ -199,7 +199,7 @@ class TestNoOptions2:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
+            "moban.jinja2.engine.verify_the_existence_of_directories"
         )
         self.patcher2 = patch(
             "moban.engine_factory.verify_the_existence_of_directories"
@@ -207,7 +207,7 @@ class TestNoOptions2:
         self.patcher1.start()
         self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
@@ -238,7 +238,7 @@ class TestCustomMobanFile:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
+            "moban.jinja2.engine.verify_the_existence_of_directories"
         )
         self.patcher2 = patch(
             "moban.engine_factory.verify_the_existence_of_directories"
@@ -246,7 +246,7 @@ class TestCustomMobanFile:
         self.patcher1.start()
         self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban", "-m", self.config_file]
         with patch.object(sys, "argv", test_args):
@@ -267,7 +267,7 @@ class TestCustomMobanFile:
         os.unlink(self.data_file)
 
 
-@patch("moban.engine.verify_the_existence_of_directories")
+@patch("moban.jinja2.engine.verify_the_existence_of_directories")
 def test_duplicated_targets_in_moban_file(fake_verify):
     config_file = "duplicated.moban.yml"
     copyfile(os.path.join("tests", "fixtures", config_file), ".moban.yml")
@@ -282,7 +282,7 @@ class TestInvalidMobanFile:
         self.config_file = ".moban.yml"
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_no_configuration(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("")
@@ -291,7 +291,7 @@ class TestInvalidMobanFile:
             main()
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_no_configuration_2(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("not: related")
@@ -300,7 +300,7 @@ class TestInvalidMobanFile:
             main()
 
     @raises(SystemExit)
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_no_targets(self, fake_template_doer):
         with open(self.config_file, "w") as f:
             f.write("configuration: test")
@@ -322,7 +322,7 @@ class TestComplexOptions:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.engine.verify_the_existence_of_directories"
+            "moban.jinja2.engine.verify_the_existence_of_directories"
         )
         self.patcher2 = patch(
             "moban.engine_factory.verify_the_existence_of_directories"
@@ -330,7 +330,7 @@ class TestComplexOptions:
         self.patcher1.start()
         self.patcher2.start()
 
-    @patch("moban.engine.Engine.render_to_files")
+    @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
@@ -357,7 +357,7 @@ class TestTemplateTypeOption:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
 
-    @patch("moban.engine.Engine.render_to_file")
+    @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_mako_option(self, fake_template_doer):
         test_args = ["moban", "-t", "a.mako"]
         with patch.object(sys, "argv", test_args):
