+++ tests/test_text_filter.py
@@ -1,5 +1,5 @@
 from nose.tools import eq_
-from moban.filters.text import split_length
+from moban.jinja2.filters.text import split_length
 
 
 def test_split_length():
